Move "sys types" utils to own module.
