+++ src/client/common/variables/sysTypes.ts
@@ -4,127 +4,9 @@
  *--------------------------------------------------------------------------------------------*/
 'use strict';
 
-/* tslint:disable:rule1 no-any no-unnecessary-callback-wrapper jsdoc-format no-for-in prefer-const no-increment-decrement */
+// tslint:disable:no-any no-increment-decrement
 
-const _typeof = {
-    number: 'number',
-    string: 'string',
-    undefined: 'undefined',
-    object: 'object',
-    function: 'function'
-};
-
-/**
- * @returns whether the provided parameter is a JavaScript Array or not.
- */
-export function isArray(array: any): array is any[] {
-    if (Array.isArray) {
-        return Array.isArray(array);
-    }
-
-    if (array && typeof (array.length) === _typeof.number && array.constructor === Array) {
-        return true;
-    }
-
-    return false;
-}
-
-/**
- * @returns whether the provided parameter is a JavaScript String or not.
- */
-export function isString(str: any): str is string {
-    if (typeof (str) === _typeof.string || str instanceof String) {
-        return true;
-    }
-
-    return false;
-}
-
-/**
- * @returns whether the provided parameter is a JavaScript Array and each element in the array is a string.
- */
-export function isStringArray(value: any): value is string[] {
-    return isArray(value) && (<any[]>value).every(elem => isString(elem));
-}
-
-/**
- *
- * @returns whether the provided parameter is of type `object` but **not**
- *	`null`, an `array`, a `regexp`, nor a `date`.
- */
-export function isObject(obj: any): obj is any {
-    return typeof obj === _typeof.object
-        && obj !== null
-        && !Array.isArray(obj)
-        && !(obj instanceof RegExp)
-        && !(obj instanceof Date);
-}
-
-/**
- * In **contrast** to just checking `typeof` this will return `false` for `NaN`.
- * @returns whether the provided parameter is a JavaScript Number or not.
- */
-export function isNumber(obj: any): obj is number {
-    if ((typeof (obj) === _typeof.number || obj instanceof Number) && !isNaN(obj)) {
-        return true;
-    }
-
-    return false;
-}
-
-/**
- * @returns whether the provided parameter is a JavaScript Boolean or not.
- */
-export function isBoolean(obj: any): obj is boolean {
-    return obj === true || obj === false;
-}
-
-/**
- * @returns whether the provided parameter is undefined.
- */
-export function isUndefined(obj: any): boolean {
-    return typeof (obj) === _typeof.undefined;
-}
-
-/**
- * @returns whether the provided parameter is undefined or null.
- */
-export function isUndefinedOrNull(obj: any): boolean {
-    return isUndefined(obj) || obj === null;
-}
-
-const hasOwnProperty = Object.prototype.hasOwnProperty;
-
-/**
- * @returns whether the provided parameter is an empty JavaScript Object or not.
- */
-export function isEmptyObject(obj: any): obj is any {
-    if (!isObject(obj)) {
-        return false;
-    }
-
-    for (let key in obj) {
-        if (hasOwnProperty.call(obj, key)) {
-            return false;
-        }
-    }
-
-    return true;
-}
-
-/**
- * @returns whether the provided parameter is a JavaScript Function or not.
- */
-export function isFunction(obj: any): obj is Function {
-    return typeof obj === _typeof.function;
-}
-
-/**
- * @returns whether the provided parameters is are JavaScript Function or not.
- */
-export function areFunctions(...objects: any[]): boolean {
-    return objects && objects.length > 0 && objects.every(isFunction);
-}
+import { isFunction, isString } from '../utils/sysTypes';
 
 export type TypeConstraint = string | Function;
